+++ lib/TH/THAtomic.c
@@ -23,7 +23,7 @@ void THAtomicSet(int volatile *a, int newvalue)
 #if defined(USE_C11_ATOMICS)
   atomic_store(a, newvalue);
 #elif defined(USE_MSC_ATOMICS)
-  _InterlockedExchange((long*)a, newvalue);
+  _InterlockedExchange((size_t*)a, newvalue);
 #elif defined(USE_GCC_ATOMICS)
   __sync_lock_test_and_set(a, newvalue);
 #else
@@ -52,7 +52,7 @@ int THAtomicAdd(int volatile *a, int value)
 #if defined(USE_C11_ATOMICS)
   return atomic_fetch_add(a, value);
 #elif defined(USE_MSC_ATOMICS)
-  return _InterlockedExchangeAdd((long*)a, value);
+  return _InterlockedExchangeAdd((size_t*)a, value);
 #elif defined(USE_GCC_ATOMICS)
   return __sync_fetch_and_add(a, value);
 #else
